Makefile: Remove explicit CC assignment
The explicit assignment breaks build systems that cross compile this
package and need to override CC to the cross compiler toolchain.
Instead, delete this assignment all together.  The make tool will
default to using the "cc" compiler on most systems which is typically a
symlink to gcc anyways.

This allows the environment variables to safely override the compiler in
a standardized way.

Signed-off-by: Kyle Manna <kyle.manna@fuel7.com>
Signed-off-by: Stephen Kitt <steve@sk2.org>