Avoid busy looping when no device is attached
Josh Bowman <jkbo@users.sourceforge.net> pointed out that the change
in 5b2e145a means that read() returning without flagging an error
causes inputattach to loop. When run on a serial port with no attached
device, this loop ends up becoming a busy loop after 400 iterations or
so.

With an N_MOUSE line discipline, the read() call is just supposed to
block indefinitely until an error occurs or the device goes away; it's
not supposed to be called in a loop. This change fixes the use of
read() and only loops for errors calling for a retry.