Add mouseman4 mode, to support 4-button Logitech mice.
The Logitech Trackman Marble FX has 4 buttons. When connected via a
serial port sermouse.c will report this as BTN_SIDE, but the capability
is only announced if extra & 0x02 is set. Without the capability,
the BTN_SIDE events get filtered.

Add an extra mode that allows use of the 4th button.

Signed-off-by: Jamie Lentin <jm@lentin.co.uk>